Moving packages from parent package to repodetails package.
