+++ app/src/main/java/br/com/luisfernandez/github/client/koin/ModelModule.kt
@@ -1,9 +1,9 @@
 package br.com.luisfernandez.github.client.koin
 
-import br.com.luisfernandez.github.client.issuelist.IssueListModel
-import br.com.luisfernandez.github.client.issuelist.IssueListModelImpl
-import br.com.luisfernandez.github.client.pullrequest.PullRequestModel
-import br.com.luisfernandez.github.client.pullrequest.PullRequestModelImpl
+import br.com.luisfernandez.github.client.repodetails.issuelist.IssueListModel
+import br.com.luisfernandez.github.client.repodetails.issuelist.IssueListModelImpl
+import br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestModel
+import br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestModelImpl
 import br.com.luisfernandez.github.client.repolist.RepoListModel
 import br.com.luisfernandez.github.client.repolist.RepoListModelImpl
 import org.koin.dsl.module.module
@@ -16,8 +16,8 @@ val modelModule = module {
         )
     }
 
-    single<PullRequestModel> {
-        PullRequestModelImpl(
+    single<br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestModel> {
+        br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestModelImpl(
                 get()
         )
     }
