+++ app/src/main/java/br/com/luisfernandez/github/client/koin/PresenterModule.kt
@@ -1,7 +1,7 @@
 package br.com.luisfernandez.github.client.koin
 
-import br.com.luisfernandez.github.client.pullrequest.PullRequestPresenter
-import br.com.luisfernandez.github.client.pullrequest.PullRequestPresenterImpl
+import br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestPresenter
+import br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestPresenterImpl
 import br.com.luisfernandez.github.client.repolist.RepoListPresenter
 import br.com.luisfernandez.github.client.repolist.RepoListPresenterImpl
 import org.koin.dsl.module.module
@@ -11,7 +11,7 @@ val presenterModule = module {
         RepoListPresenterImpl(get())
     }
 
-    single<PullRequestPresenter> {
-        PullRequestPresenterImpl(get())
+    single<br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestPresenter> {
+        br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestPresenterImpl(get())
     }
 }
