+++ app/src/main/java/br/com/luisfernandez/github/client/koin/ViewModelModule.kt
@@ -2,8 +2,8 @@
 
 package br.com.luisfernandez.github.client.koin
 
-import br.com.luisfernandez.github.client.issuelist.IssueListViewModel
-import br.com.luisfernandez.github.client.pullrequest.PullRequestViewModel
+import br.com.luisfernandez.github.client.repodetails.issuelist.IssueListViewModel
+import br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestViewModel
 import br.com.luisfernandez.github.client.repolist.RepoListViewModel
 import org.koin.android.viewmodel.ext.koin.viewModel
 import org.koin.dsl.module.module
@@ -16,8 +16,8 @@ val viewModelModule = module {
         )
     }
 
-    viewModel<PullRequestViewModel> {
-        PullRequestViewModel(
+    viewModel<br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestViewModel> {
+        br.com.luisfernandez.github.client.repodetails.pullrequest.PullRequestViewModel(
                 get()
         )
     }
