--- app/src/main/java/br/com/luisfernandez/github/client/pullrequest/PullRequestPresenter.kt
@@ -1,9 +1,0 @@
-package br.com.luisfernandez.github.client.pullrequest
-
-/**
- * Created by luisfernandez on 12/05/18.
- */
-interface PullRequestPresenter {
-    fun loadPullRequestList(owner: String, repoName: String)
-    fun inject(pullRequestView: PullRequestListView)
-}
