+++ app/src/main/java/br/com/luisfernandez/github/client/repodetails/issuelist/IssueListActivity.kt
@@ -1,18 +1,15 @@
-package br.com.luisfernandez.github.client.pullrequest
+package br.com.luisfernandez.github.client.repodetails.issuelist
 
 import android.annotation.SuppressLint
 import android.arch.lifecycle.Observer
-import android.content.Intent
-import android.net.Uri
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
-import br.com.luisfernandez.github.client.OnItemClickListener
 import br.com.luisfernandez.github.client.R
 import br.com.luisfernandez.github.client.extensions.setGone
 import br.com.luisfernandez.github.client.extensions.setVisible
-import br.com.luisfernandez.github.client.http.model.ServerError
 import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
-import br.com.luisfernandez.github.client.pojo.PullRequestResponse
+import br.com.luisfernandez.github.client.http.model.ServerError
+import br.com.luisfernandez.github.client.pojo.IssueResponse
 import kotlinx.android.synthetic.main.activity_list.*
 import kotlinx.android.synthetic.main.view_state_empty.*
 import kotlinx.android.synthetic.main.view_state_error.*
@@ -23,8 +20,8 @@ import org.androidannotations.annotations.Extra
 import org.koin.android.viewmodel.ext.android.viewModel
 
 @SuppressLint("Registered")
-@EActivity(R.layout.activity_pull_request_list)
-class PullRequestListActivity : AppCompatActivity(), PullRequestListView {
+@EActivity(R.layout.fragment_issues_list)
+class IssueListActivity : AppCompatActivity(), IssueListView {
 
     @Extra
     lateinit var owner: String
@@ -32,7 +29,7 @@ class PullRequestListActivity : AppCompatActivity(), PullRequestListView {
     @Extra
     lateinit var repoName: String
 
-    val viewModel by viewModel<PullRequestViewModel>()
+    val viewModel by viewModel<IssueListViewModel>()
 
     @AfterViews
     fun afterViews() {
@@ -45,13 +42,13 @@ class PullRequestListActivity : AppCompatActivity(), PullRequestListView {
 
         setupViewModel()
 
-        viewModel.loadPullRequestList(owner, repoName)
+        viewModel.loadIssueList(owner, repoName)
     }
 
     private fun setupViewModel() {
-        viewModel.listPullRequest.observe(this, Observer {
-            listPullRequest ->
-            showContent(listPullRequest!!)
+        viewModel.issueList.observe(this, Observer {
+            issueList ->
+            showContent(issueList!!)
         })
 
         viewModel.serverError.observe(this, Observer {
@@ -86,7 +83,7 @@ class PullRequestListActivity : AppCompatActivity(), PullRequestListView {
 
         buttonRetry.setOnClickListener { _ ->
             showLoading()
-            viewModel.loadPullRequestList(owner, repoName)
+            viewModel.loadIssueList(owner, repoName)
         }
     }
 
@@ -111,7 +108,7 @@ class PullRequestListActivity : AppCompatActivity(), PullRequestListView {
         recyclerView.setGone()
     }
 
-    override fun showContent(content: List<PullRequestResponse>) {
+    override fun showContent(content: List<IssueResponse>) {
         if (content.isEmpty()) {
             showEmpty()
         } else {
@@ -120,28 +117,13 @@ class PullRequestListActivity : AppCompatActivity(), PullRequestListView {
             layoutError.setGone()
             recyclerView.setVisible()
 
-            recyclerView.adapter = this.getPullRequestListAdapter(content)
+            recyclerView.adapter = this.getIssueListAdapter(content)
         }
     }
 
-    private fun getPullRequestListAdapter(content: List<PullRequestResponse>): PullRequestListAdapter {
-        return PullRequestListAdapter(
-                content as ArrayList<PullRequestResponse>,
-                getOnItemClickListener()
+    private fun getIssueListAdapter(content: List<IssueResponse>): IssueListAdapter {
+        return IssueListAdapter(
+                content as ArrayList<IssueResponse>
         )
-    }
-
-    private fun getOnItemClickListener(): OnItemClickListener<PullRequestResponse> {
-        return object : OnItemClickListener<PullRequestResponse> {
-            override fun onItemClick(type: PullRequestResponse) {
-                openOnBrowser(type.htmlUrl)
-            }
-        }
-    }
-
-    private fun openOnBrowser(url: String) {
-        val intent = Intent(Intent.ACTION_VIEW)
-        intent.data = Uri.parse(url)
-        startActivity(intent)
     }
 }
