+++ app/src/main/java/br/com/luisfernandez/github/client/repodetails/issuelist/IssueListModelImpl.kt
@@ -1,4 +1,4 @@
-package br.com.luisfernandez.github.client.issuelist
+package br.com.luisfernandez.github.client.repodetails.issuelist
 
 import br.com.luisfernandez.github.client.http.GitHubService
 import br.com.luisfernandez.github.client.pojo.IssueResponse
