+++ app/src/main/res/layout/fragment_issues_list.xml
@@ -3,7 +3,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:animateLayoutChanges="true"
-    tools:context=".repodetails.IssuesListFragment">
+    tools:context=".repodetails.issuelist.IssuesListFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView"
